requirements.txt
flask
slack_bolt
slack_sdk
openai
PyMuPDF
python-dotenv
requests
beautifulsoup4
gunicorn
supabase
